Geofencing with leaflet example
Code snippet of what it might be like to use getbounds call in leaflet
on a circle to mimic a geofence.

One marker shows last recorded location, and the other is draggable
allowing for the red "geofenced" area to be drawn around that marker.
The blue circle is centered on the location marker and the radius
changes with the last logged accuracy.