Zoom Responsive, joyride added, bug fixed
Zoom level on POI now changes based on window width, zoomed further out
for smaller screen sizes.

Joyride added to 640px or less to inform user to press play and to
click on menu.  The padding and position needs to be fixed.  This may
change to be only responsive to media type and not screen size.

Fixed a bug on the nav bar locations selection.